- Fixed critical bugs with new 'doubles' support.

+++ protomuck/src/p_stack.c
@@ -22,7 +22,7 @@ struct inst *oper1, *oper2, *oper3, *oper4;
 struct inst temp1, temp2, temp3;
 struct inst arrayvar;
 int result, tmp;
-float fresult;
+double fresult;
 dbref ref;
 char buf[BUFFER_LEN];
